requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
langchain-openai>=0.0.1
anthropic>=0.7.0
